PPO/requirements.txt
torch>=2.1.0
numpy>=1.24.0
gymnasium[classic-control]>=0.29.0
tqdm>=4.66.0
matplotlib>=3.7.0
pyyaml>=6.0
pytest>=7.4.0
pandas>=2.0.0
